+++ app/src/main/java/com/example/scamdetectorapp/data/model/AiCheckRequest.kt
@@ -1,0 +1,5 @@
+package com.example.scamdetectorapp.data.model
+
+data class AiCheckRequest(
+    val text: String
+)
