+++ app/src/main/java/com/example/scamdetectorapp/data/model/AiCheckResult.kt
@@ -1,0 +1,7 @@
+package com.example.scamdetectorapp.data.model
+
+data class AiCheckResult(
+    val riskLevel: String? = null,
+    val description: String? = null,
+    val suggestion: String? = null
+)
