+++ app/src/main/java/com/example/scamdetectorapp/data/model/UrlCheckResult.kt
@@ -1,0 +1,7 @@
+package com.example.scamdetectorapp.data.model
+
+data class UrlCheckResult(
+    val riskLevel: String? = null,
+    val description: String? = null,
+    val threatType: String? = null
+)
