+++ app/src/main/java/com/example/scamdetectorapp/data/remote/RetrofitClient.kt
@@ -1,0 +1,34 @@
+package com.example.scamdetectorapp.data.remote
+
+import com.example.scamdetectorapp.BuildConfig
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+
+object RetrofitClient {
+    private const val BASE_URL = "https://antifraud-gateway.lyc-dev.workers.dev/"
+
+    private val okHttpClient = OkHttpClient.Builder()
+        .addInterceptor { chain ->
+            val original = chain.request()
+            val request = original.newBuilder()
+                .header("x-api-key", BuildConfig.CLOUDFLARE_API_KEY)
+                .method(original.method, original.body)
+                .build()
+            chain.proceed(request)
+        }
+        .connectTimeout(30, TimeUnit.SECONDS)
+        .readTimeout(30, TimeUnit.SECONDS)
+        .writeTimeout(30, TimeUnit.SECONDS)
+        .build()
+
+    val instance: AntiFraudApi by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .client(okHttpClient)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(AntiFraudApi::class.java)
+    }
+}
