+++ app/src/main/java/com/example/scamdetectorapp/data/repository/AntiFraudRepository.kt
@@ -1,0 +1,62 @@
+package com.example.scamdetectorapp.data.repository
+
+import com.example.scamdetectorapp.data.model.*
+import com.example.scamdetectorapp.data.remote.RetrofitClient
+import com.example.scamdetectorapp.domain.model.DetectionMode
+import com.example.scamdetectorapp.domain.model.ScanResult
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+class AntiFraudRepository {
+    private val api = RetrofitClient.instance
+
+    suspend fun scan(mode: DetectionMode, input: String): Result<ScanResult> = withContext(Dispatchers.IO) {
+        try {
+            val result = when (mode) {
+                DetectionMode.PHONE -> {
+                    val response = api.getData(phoneNumber = input)
+                    if (response.success) {
+                        val data = response.data
+                        ScanResult(
+                            riskLevel = data?.riskLevel ?: "UNKNOWN",
+                            description = data?.description
+                        )
+                    } else throwApiException(response.version)
+                }
+                DetectionMode.URL -> {
+                    var urlToCheck = input.trim()
+                    if (!urlToCheck.startsWith("http://") && !urlToCheck.startsWith("https://")) {
+                        urlToCheck = "https://$urlToCheck"
+                    }
+                    val response = api.getUrlCheck(url = urlToCheck)
+                    if (response.success) {
+                        val data = response.data
+                        ScanResult(
+                            riskLevel = data?.riskLevel ?: "UNKNOWN",
+                            description = data?.description,
+                            threatType = data?.threatType
+                        )
+                    } else throwApiException(response.version)
+                }
+                DetectionMode.TEXT -> {
+                    val response = api.postAiCheck(body = AiCheckRequest(text = input))
+                    if (response.success) {
+                        val data = response.data
+                        ScanResult(
+                            riskLevel = data?.riskLevel ?: "UNKNOWN",
+                            description = data?.description,
+                            suggestion = data?.suggestion
+                        )
+                    } else throwApiException(response.version)
+                }
+            }
+            Result.success(result)
+        } catch (e: Exception) {
+            Result.failure(e)
+        }
+    }
+
+    private fun throwApiException(version: String): Nothing {
+        throw Exception("API 回傳失敗: $version")
+    }
+}
