+++ app/src/main/java/com/example/scamdetectorapp/domain/model/DetectionMode.kt
@@ -1,0 +1,5 @@
+package com.example.scamdetectorapp.domain.model
+
+enum class DetectionMode {
+    URL, PHONE, TEXT
+}
