+++ app/src/main/java/com/example/scamdetectorapp/presentation/model/ScanUiModel.kt
@@ -1,0 +1,8 @@
+package com.example.scamdetectorapp.presentation.model
+
+data class ScanUiModel(
+    val isSafe: Boolean,
+    val score: Int,
+    val title: String,
+    val reasons: List<String>
+)
