+++ app/src/main/java/com/example/scamdetectorapp/presentation/screens/splash/SplashScreen.kt
@@ -1,0 +1,62 @@
+package com.example.scamdetectorapp.presentation.screens.splash
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Shield
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.scamdetectorapp.R
+import kotlinx.coroutines.delay
+
+@Composable
+fun SplashScreen(onFinished: () -> Unit) {
+    val darkBg = colorResource(R.color.scam_background)
+    val primaryColor = colorResource(R.color.scam_primary)
+    val textColor = colorResource(R.color.scam_text_white)
+    val textGrey = colorResource(R.color.scam_text_grey)
+
+    LaunchedEffect(Unit) {
+        delay(2000)
+        onFinished()
+    }
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Brush.verticalGradient(listOf(Color(0xFF0F172A), darkBg))),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            Icon(
+                imageVector = Icons.Outlined.Shield,
+                contentDescription = "Logo",
+                tint = primaryColor,
+                modifier = Modifier.size(80.dp)
+            )
+            Spacer(modifier = Modifier.height(16.dp))
+            Text(
+                "SCAM GUARD",
+                fontSize = 28.sp,
+                fontWeight = FontWeight.Bold,
+                letterSpacing = 2.sp,
+                color = textColor
+            )
+            Spacer(modifier = Modifier.height(8.dp))
+            Text("Real Time Protection", color = textGrey, fontSize = 14.sp)
+            Spacer(modifier = Modifier.height(48.dp))
+            CircularProgressIndicator(color = primaryColor, modifier = Modifier.size(32.dp))
+        }
+    }
+}
